feat(deploy): wrangler-applied D1 migrations everywhere + runtime self-seed
Deploy-button clones shipped an unmigrated, unseeded DB: nothing ran
migrations (Cloudflare's default deploy command is a bare wrangler
deploy) and seeding needed CI-only HTTP credentials.

- Add a top-level 'deploy' script (the button/Workers Builds deploy
  command): flatten -> wrangler d1 migrations apply DB --remote ->
  wrangler deploy. scripts/flatten-migrations.ts renders drizzle-kit's
  nested <dir>/migration.sql into flat gitignored
  drizzle/migrations-wrangler/*.sql, which is the only layout wrangler
  reads.
- Prod CI migrates the same way via db:migrate:prd inside cf:deploy:prd;
  the drizzle-kit d1-http migrate step (the #612 single-body vector) and
  both CI seed steps are gone. Preview CI applies migrations after the
  config patch (which now preserves migrations_dir).
- Runtime self-seed: seed logic moves to
  src/lib/db/seed-system-templates.ts (shared with scripts/seed.ts).
  The worker seeds on first request, gated by a SHA-256 of the template
  definitions stored in the new app_metadata table (timestamps excluded
  from the hash - templates stamp new Date() at module load), guarded by
  a TTL'd lock row so exactly one isolate runs a stale-hash sync, and
  skipped entirely for /r2/ media requests.
- Fix two latent schema-drift landmines (#898): restore the
  creditBatches/teamBillingSettings exports the knip sweep stripped
  (drizzle-kit treats non-exported tables as dropped) and pin
  sequences.videoModel's SQL default back to kling_v3_pro with
  $defaultFn supplying the grok default app-side (the unmigrated
  default change would have generated a CASCADE-unsafe rebuild).
- Document the velro-prd -> openstory-prd cutover runbook (D1 has no
  rename; recreate-from-migrations gives native d1_migrations tracking
  with no baselining) in docs/deployment/cloudflare.md.

Closes #897
