+++ .github/workflows/deploy-cloudflare.yml
@@ -45,21 +45,10 @@ jobs:
       - name: Install dependencies
         run: bun install
 
-      - name: Migrate D1 (production)
-        run: bun drizzle-kit migrate --config=drizzle.config.d1.ts
-        env:
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          CLOUDFLARE_D1_DATABASE_ID: d6a35f64-abb0-4a34-a092-dc20b29f5812
-
-      - name: Seed D1 (production)
-        run: bun run db:seed:d1
-        env:
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          CLOUDFLARE_D1_DATABASE_ID: d6a35f64-abb0-4a34-a092-dc20b29f5812
-          VITE_R2_PUBLIC_ASSETS_DOMAIN: ${{ vars.VITE_R2_PUBLIC_ASSETS_DOMAIN }}
-
+      # Migrations are applied inside cf:deploy:prd (wrangler d1 migrations
+      # apply DB --env=production --remote, after the build and before the
+      # deploy). Seeding happens at runtime: the worker self-seeds system
+      # templates on first request (src/server.ts). See #897.
       - name: Deploy to production
         run: bun cf:deploy:prd
         env:
@@ -67,6 +56,9 @@ jobs:
           CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           VITE_PUBLIC_POSTHOG_PROJECT_TOKEN: ${{ secrets.VITE_PUBLIC_POSTHOG_PROJECT_TOKEN }}
           VITE_PUBLIC_POSTHOG_HOST: ${{ vars.VITE_PUBLIC_POSTHOG_HOST }}
+          # Baked into sheet/preview URLs by the runtime self-seed
+          # (import.meta.env, build-time).
+          VITE_R2_PUBLIC_ASSETS_DOMAIN: ${{ vars.VITE_R2_PUBLIC_ASSETS_DOMAIN }}
 
   # ── PR preview deploy ─────────────────────────────────────────────
   deploy-preview:
@@ -134,21 +126,6 @@ jobs:
             echo "database-id=$uuid" >> "$GITHUB_OUTPUT"
           fi
 
-      - name: Migrate D1 (preview)
-        run: bun drizzle-kit migrate --config=drizzle.config.d1.ts
-        env:
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          CLOUDFLARE_D1_DATABASE_ID: ${{ steps.d1.outputs.database-id }}
-
-      - name: Seed D1 (preview)
-        run: bun run db:seed:d1
-        env:
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          CLOUDFLARE_D1_DATABASE_ID: ${{ steps.d1.outputs.database-id }}
-          VITE_R2_PUBLIC_ASSETS_DOMAIN: ${{ vars.VITE_R2_PUBLIC_ASSETS_DOMAIN }}
-
       - name: Patch wrangler config for preview
         env:
           PR_NUMBER: ${{ github.event.pull_request.number }}
@@ -163,7 +140,8 @@ jobs:
             obj.d1_databases = [{
               binding: 'DB',
               database_name: 'openstory-pr-${PR_NUMBER}',
-              database_id: '${D1_DATABASE_ID}'
+              database_id: '${D1_DATABASE_ID}',
+              migrations_dir: 'drizzle/migrations-wrangler'
             }];
             obj.r2_buckets = [
               { bucket_name: 'openstory-public-assets', binding: 'R2_PUBLIC_ASSETS_BUCKET', remote: true },
@@ -198,12 +176,26 @@ jobs:
           echo "Patched wrangler.jsonc:"
           cat wrangler.jsonc | head -30
 
+      # Runs after the patch so wrangler resolves the PR's own D1 (the
+      # binding's database_id was just rewritten). Seeding happens at
+      # runtime via the worker's self-seed (src/server.ts). See #897.
+      - name: Migrate D1 (preview)
+        env:
+          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
+          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
+        run: |
+          bun scripts/flatten-migrations.ts
+          bunx wrangler d1 migrations apply DB --remote
+
       - name: Build
         env:
           CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
           CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           VITE_PUBLIC_POSTHOG_PROJECT_TOKEN: ${{ secrets.VITE_PUBLIC_POSTHOG_PROJECT_TOKEN }}
           VITE_PUBLIC_POSTHOG_HOST: ${{ vars.VITE_PUBLIC_POSTHOG_HOST }}
+          # Baked into sheet/preview URLs by the runtime self-seed
+          # (import.meta.env, build-time).
+          VITE_R2_PUBLIC_ASSETS_DOMAIN: ${{ vars.VITE_R2_PUBLIC_ASSETS_DOMAIN }}
         run: |
           bunx wrangler types
           bun run build
