+++ .gitignore
@@ -85,6 +85,9 @@ conductor.json
 .wrangler/
 # Slimmed wrangler config generated by scripts/local-platform-proxy.ts (#859)
 wrangler.local-proxy.*.jsonc
+# Flat *.sql rendering of drizzle/migrations for `wrangler d1 migrations apply`,
+# rebuilt on demand by scripts/flatten-migrations.ts (#897)
+drizzle/migrations-wrangler/
 
 # output
 .output/
